+++ src/wdio/services/AxeService.js
@@ -12,12 +12,13 @@ export default class AxeService {
   // eslint-disable-next-line class-methods-use-this
   before() {
     browser.addCommand('axe', (options = {}) => {
-      // Conditionally inject axe. This allows consumers to inject it themselves
-      // in the test examples which would slightly speed up test runs.
       const axeConfig = {
         ...SERVICE_DEFAULTS.axe,
         ...(browser.options.axe || {}),
       };
+
+      // Conditionally inject axe. This allows consumers to inject it themselves
+      // in the test examples which would slightly speed up test runs.
       if (axeConfig.inject) {
         if (browser.execute('return window.axe === undefined;')) {
           if (!axeCoreSrc) {
@@ -34,35 +35,44 @@ export default class AxeService {
         }
       }
 
-      const currentViewportSize = browser.getViewportSize();
       // use current viewport if none specified
-      const specifiedViewports = options.viewports || [currentViewportSize];
+      const specifiedViewports = options.viewports;
       const axeOptions = {
         runOnly: options.runOnly,
         rules: options.rules,
       };
 
-      // Get accessibility results for each viewport size
-      const results = specifiedViewports.map((viewport) => {
-        browser.setViewportSize(viewport);
+      const runAxeTest = (wdioContext, wdioOptions) => {
         // Avoid arrow callback syntax as this function is injected into the browser
         // eslint-disable-next-line func-names, prefer-arrow-callback
-        return browser.executeAsync(function (context, opts, done) {
-          // eslint-disable-next-line func-names, prefer-arrow-callback
+        const axeResult = browser.executeAsync(function (context, opts, done) {
+        // eslint-disable-next-line func-names, prefer-arrow-callback
           axe.run(context || document, opts, function (error, result) {
-            done({
-              // eslint-disable-next-line object-shorthand
-              error: error,
-              // eslint-disable-next-line object-shorthand
-              result: result,
-            });
+            /* eslint-disable-next-line object-shorthand */
+            done({ error: error, result: result });
           });
-        }, options.context, axeOptions).value;
-      });
+        }, wdioContext, wdioOptions);
 
-      // set viewport back
-      browser.setViewportSize(currentViewportSize);
-      return results;
+        return axeResult.value;
+      };
+
+      let results;
+      if (specifiedViewports) {
+        const currentViewportSize = browser.getViewportSize();
+
+        // Get accessibility results for each viewport size
+        results = specifiedViewports.map((viewport) => {
+          browser.setViewportSize(viewport);
+          return runAxeTest(options.context, axeOptions);
+        });
+
+        // set viewport back
+        browser.setViewportSize(currentViewportSize);
+
+        return results;
+      }
+
+      return [runAxeTest(options.context, axeOptions)];
     });
   }
 }