+++ config/webpack/webpack.config.js
@@ -8,6 +8,7 @@ const CleanPlugin = require('clean-webpack-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
 const browserslist = require('browserslist-config-terra');
 const merge = require('webpack-merge');
+const DuplicatePackageCheckerPlugin = require('duplicate-package-checker-webpack-plugin');
 const aggregateTranslations = require('../../scripts/aggregate-translations/aggregate-translations');
 
 const webpackConfig = (options, env, argv) => {
@@ -82,6 +83,9 @@ const webpackConfig = (options, env, argv) => {
         plugins: [
           PostCSSCustomProperties({ preserve: true }),
         ],
+      }),
+      new DuplicatePackageCheckerPlugin({
+        showHelp: false,
       }),
     ],
     resolve: {