+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update the `connectionRetryCount` to 1 in the default wdio configuraiton
+
 ### Added
 * Added format option to aggregate-translations to output modern syntax