+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update the `connectionRetryCount` to 1 in the default wdio configuraiton
+
+### Added
+* Added format option to aggregate-translations to output modern syntax
 
 4.22.0 - (January 2, 2019)
 ----------