+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 
 ### Added
 * Added format option to aggregate-translations to output modern syntax
+* Added the Duplicate Package Checker Plugin to the default webpack config. The duplicate package checker plugin will warn if duplicate packages (different versions) are included in your webpack bundle.
 
 4.22.0 - (January 2, 2019)
 ----------