+++ CHANGELOG.md
@@ -3,12 +3,15 @@ Changelog
 
 Unreleased
 ----------
-### Changed
-* Update the `connectionRetryCount` to 1 in the default wdio configuraiton
-
 ### Added
 * Added format option to aggregate-translations to output modern syntax
 * Added the Duplicate Package Checker Plugin to the default webpack config. The duplicate package checker plugin will warn if duplicate packages (different versions) are included in your webpack bundle.
+
+### Changed
+* Update the `connectionRetryCount` to 1 in the default wdio configuraiton
+
+### Fixed
+* Adjust AxeService implementation to only resize the viewport if viewport options are passed.
 
 4.22.0 - (January 2, 2019)
 ----------