+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Unreleased
 
 ### Changed
 * Update the `connectionRetryCount` to 1 in the default wdio configuraiton
+* Added the Duplicate Package Checker Plugin to the default webpack config. The duplicate package checker plugin will warn if duplicate packages (different versions) are included in your webpack bundle.
+
+### Fixed
+* Adjust AxeService implementation to only resize the viewport if viewport options are passed.
 
 4.22.0 - (January 2, 2019)
 ----------